Refactored the Top and Bottom App Bars, removing them from the HomeScreen to be their own separate composables.
Added a navigation drawer to aid in navigating the app. This retrieves basic user information at the moment but can be easily expanded upon. One issue that arose after implementing the nav drawer though is that it has skewed the layout of the login, registration and edit routes pages. I moved the scaffold to a higher level, the main activity thinking that would resolve it but it hasn't so I left it on the Home Screen. This is because I need to condition displaying the app bars and if the scaffold is not rectifying the issue it's not worth changing.